Torus Object implemented, clean up some inconsistencies, clean up RenderUtil.java, support for rotation and floating number coordinates

+++ src/main/java/de/srendi/advancedperipherals/common/smartglasses/modules/overlay/objects/two_dim/CircleObject.java
@@ -58,12 +58,12 @@ public class CircleObject extends RenderableObject {
         int color = buffer.readInt();
         float opacity = buffer.readFloat();
 
-        int x = buffer.readInt();
-        int y = buffer.readInt();
-        int z = buffer.readInt();
-        int maxX = buffer.readInt();
-        int maxY = buffer.readInt();
-        int maxZ = buffer.readInt();
+        float x = buffer.readFloat();
+        float y = buffer.readFloat();
+        float z = buffer.readFloat();
+        float maxX = buffer.readFloat();
+        float maxY = buffer.readFloat();
+        float maxZ = buffer.readFloat();
         int radius = buffer.readInt();
 
         CircleObject clientObject = new CircleObject(player);
