+++ src/main/java/de/srendi/advancedperipherals/common/smartglasses/modules/overlay/objects/two_dim/ItemObject.java
@@ -54,12 +54,12 @@ public class ItemObject extends RenderableObject {
         int color = buffer.readInt();
         float opacity = buffer.readFloat();
 
-        int x = buffer.readInt();
-        int y = buffer.readInt();
-        int z = buffer.readInt();
-        int maxX = buffer.readInt();
-        int maxY = buffer.readInt();
-        int maxZ = buffer.readInt();
+        float x = buffer.readFloat();
+        float y = buffer.readFloat();
+        float z = buffer.readFloat();
+        float maxX = buffer.readFloat();
+        float maxY = buffer.readFloat();
+        float maxZ = buffer.readFloat();
         String item = buffer.readUtf();
 
         ItemObject clientObject = new ItemObject(player);
